Adjusted popular archives colors. Resolved PR comments.
