+++ .idea/misc.xml
@@ -12,6 +12,9 @@
         <entry key="app/src/main/res/drawable/googlepay_button_overlay.xml" value="0.2685" />
         <entry key="app/src/main/res/drawable/ic_delete_white.xml" value="0.4197916666666667" />
         <entry key="app/src/main/res/drawable/ic_edit.xml" value="0.4197916666666667" />
+        <entry key="app/src/main/res/drawable/ic_link.xml" value="0.2375" />
+        <entry key="app/src/main/res/drawable/ic_link_middle_grey.xml" value="0.235" />
+        <entry key="app/src/main/res/drawable/ic_plus_white.xml" value="0.2375" />
         <entry key="app/src/main/res/drawable/login_box_background_color.xml" value="0.2585" />
         <entry key="app/src/main/res/drawable/manage_link_background_color.xml" value="0.2585" />
         <entry key="app/src/main/res/drawable/rounded_blue_border_white_background.xml" value="0.2685" />
@@ -46,7 +49,6 @@
         <entry key="app/src/main/res/layout/fragment_archives.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_archives_container.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_container.xml" value="0.4036458333333333" />
-        <entry key="app/src/main/res/layout/fragment_donation_summary.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_edit_about.xml" value="0.5820512820512821" />
         <entry key="app/src/main/res/layout/fragment_edit_archive_information.xml" value="0.3098958333333333" />
         <entry key="app/src/main/res/layout/fragment_edit_milestone.xml" value="0.2046875" />
@@ -61,9 +63,7 @@
         <entry key="app/src/main/res/layout/fragment_online_presence_list.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_public.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_public_archive.xml" value="0.3440366972477064" />
-        <entry key="app/src/main/res/layout/fragment_public_archives.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_public_files.xml" value="0.4036458333333333" />
-        <entry key="app/src/main/res/layout/fragment_public_gallery.xml" value="0.3974358974358974" />
         <entry key="app/src/main/res/layout/fragment_public_profile.xml" value="0.3098958333333333" />
         <entry key="app/src/main/res/layout/fragment_record_options.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/fragment_security.xml" value="0.4036458333333333" />
@@ -76,7 +76,7 @@
         <entry key="app/src/main/res/layout/item_milestone_underlay.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/item_online_presence.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/item_online_presence_underlay.xml" value="0.4036458333333333" />
-        <entry key="app/src/main/res/layout/item_public_archive.xml" value="0.4036458333333333" />
+        <entry key="app/src/main/res/layout/item_public_archive.xml" value="0.3974358974358974" />
         <entry key="app/src/main/res/layout/nav_main_header.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/nav_settings_header.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/menu/menu_main_drawer.xml" value="0.5820512820512821" />
