+++ app/src/main/java/org/permanent/permanent/viewmodels/PublicGalleryViewModel.kt
@@ -6,6 +6,11 @@ import android.content.ClipboardManager
 import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.google.firebase.ktx.Firebase
+import com.google.firebase.remoteconfig.FirebaseRemoteConfig
+import com.google.firebase.remoteconfig.ktx.remoteConfig
+import com.google.gson.Gson
+import com.google.gson.JsonObject
 import org.permanent.permanent.BuildConfig
 import org.permanent.permanent.R
 import org.permanent.permanent.models.Archive
@@ -31,7 +36,18 @@ class PublicGalleryViewModel(application: Application) : ObservableAndroidViewMo
     private val onPopularArchivesRetrieved = MutableLiveData<List<Archive>>()
 
     init {
-       getYourPublicArchives()
+        getYourPublicArchives()
+
+        val remoteConfig = setupRemoteConfig()
+
+        remoteConfig.fetchAndActivate().addOnCompleteListener {
+            val popularArchivesString = remoteConfig.getString(POPULAR_PUBLIC_ARCHIVES_KEY)
+            val gson = Gson()
+            val jsonObject = gson.fromJson(popularArchivesString, JsonObject::class.java)
+            val arr = jsonObject.getAsJsonArray(REMOTE_CONFIG_ARCHIVES_KEY)
+            val minimizedPopularArchives = gson.fromJson(arr, Array<Archive>::class.java).asList()
+            getPopularArchives(minimizedPopularArchives)
+        }
     }
 
     private fun getYourPublicArchives() {
@@ -48,7 +64,7 @@ class PublicGalleryViewModel(application: Application) : ObservableAndroidViewMo
                         val archives: MutableList<Archive> = ArrayList()
                         for (datum in dataList) {
                             val archive = Archive(datum.ArchiveVO)
-                            if (archive.public == 1) {
+                            if (archive.isPublic == 1) {
                                 archives.add(archive)
                             }
                         }
@@ -58,10 +74,24 @@ class PublicGalleryViewModel(application: Application) : ObservableAndroidViewMo
 
                 override fun onFailed(error: String?) {
                     isBusy.value = false
-                    showError.value = error
+                    error?.let { showError.value = it }
                 }
             })
         }
+    }
+
+    private fun setupRemoteConfig(): FirebaseRemoteConfig {
+        val remoteConfig = Firebase.remoteConfig
+
+        remoteConfig.setDefaultsAsync(R.xml.remote_config_defaults)
+
+        // FOR DEVELOPMENT PURPOSE ONLY
+        // The default minimum fetch interval is 12 hours
+//        val configSettings = remoteConfigSettings {
+//            minimumFetchIntervalInSeconds = 30
+//        }
+//        remoteConfig.setConfigSettingsAsync(configSettings)
+        return remoteConfig
     }
 
     fun getPopularArchives(minimizedPopularArchives: List<Archive>) {
@@ -75,9 +105,8 @@ class PublicGalleryViewModel(application: Application) : ObservableAndroidViewMo
                         val archives: MutableList<Archive> = ArrayList()
                         for (datum in dataList) {
                             val archive = Archive(datum.ArchiveVO)
-                            if (archive.public == 1) {
-                                archives.add(archive)
-                            }
+                            archive.isPopular = true
+                            archives.add(archive)
                         }
                         onPopularArchivesRetrieved.value = archives
                     }
@@ -85,7 +114,7 @@ class PublicGalleryViewModel(application: Application) : ObservableAndroidViewMo
 
                 override fun onFailed(error: String?) {
                     isBusy.value = false
-                    showError.value = error
+                    error?.let { showError.value = it }
                 }
             })
         }
@@ -111,4 +140,8 @@ class PublicGalleryViewModel(application: Application) : ObservableAndroidViewMo
         showMessage.value = appContext.getString(R.string.share_link_link_copied)
     }
 
+    companion object {
+        private const val POPULAR_PUBLIC_ARCHIVES_KEY = "popular_public_archives_android"
+        private const val REMOTE_CONFIG_ARCHIVES_KEY = "archives"
+    }
 }
