+++ app/src/main/res/drawable/ic_link_white.xml
@@ -5,6 +5,6 @@
     android:viewportHeight="20">
   <path
       android:pathData="m23.45,9.073 l-5.001,5.001a5.304,5.304 0,0 1,-3.758 1.555,5.305 5.305,0 0,1 -3.76,-1.555c-1.86,-1.86 -2.07,-4.85 -0.49,-6.947a0.623,0.623 0,0 1,0.874 -0.125,0.623 0.623,0 0,1 0.125,0.874 4.08,4.08 0,0 0,0.375 5.311,4.071 4.071,0 0,0 5.749,0l5,-5.001a4.069,4.069 0,0 0,0 -5.747,4.068 4.068,0 0,0 -5.747,0l-1.752,1.75a0.623,0.623 0,0 1,-0.884 0,0.623 0.623,0 0,1 0,-0.883l1.752,-1.753a5.325,5.325 0,0 1,7.517 0,5.32 5.32,0 0,1 0,7.52m-12.626,6.74a0.626,0.626 0,0 1,0 0.883L9.07,18.45A5.292,5.292 0,0 1,5.317 20a5.312,5.312 0,0 1,-3.764 -1.55,5.327 5.327,0 0,1 0,-7.518L6.556,5.9a5.322,5.322 0,0 1,7.517 0c1.86,1.86 2.072,4.848 0.493,6.947a0.623,0.623 0,0 1,-0.875 0.123,0.623 0.623,0 0,1 -0.123,-0.875c1.205,-1.572 1.043,-3.854 -0.379,-5.28a4.071,4.071 0,0 0,-5.748 0L2.44,11.818a4.073,4.073 0,0 0,0 5.75,4.071 4.071,0 0,0 5.748,0l1.752,-1.754a0.626,0.626 0,0 1,0.884 0"
-      android:fillColor="#000"
+      android:fillColor="@color/white"
       android:fillType="evenOdd"/>
 </vector>
