+++ app/src/main/res/layout/fragment_public_archive.xml
@@ -67,7 +67,7 @@
                 android:layout_width="36dp"
                 android:layout_height="wrap_content"
                 android:onClick="@{() -> viewModel.onCopyLinkBtnClick()}"
-                app:icon="@drawable/ic_link"
+                app:icon="@drawable/ic_link_white"
                 app:iconSize="16dp"
                 app:iconTint="@color/middleGrey"
                 app:layout_constraintBottom_toBottomOf="parent"
