+++ app/src/main/res/layout/item_public_archive.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:tools="http://schemas.android.com/tools"
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
+
         <import type="android.view.View" />
 
         <variable
@@ -16,7 +16,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginBottom="16dp"
-        android:background="@color/colorPrimary">
+        android:background="@{archive.isPopular? @color/lighterGrey : @color/colorPrimary}">
 
         <ImageView
             android:id="@+id/ivGalleryThumb"
@@ -39,7 +39,7 @@
             android:fontFamily="@font/open_sans_bold"
             android:maxLines="2"
             android:text="@{archive.fullName}"
-            android:textColor="@color/white"
+            android:textColor="@{archive.isPopular? @color/colorPrimary : @color/white}"
             android:textSize="18sp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.0"
@@ -56,19 +56,20 @@
             android:text="@{archive.accessRoleText}"
             android:textColor="@color/white"
             android:textSize="12sp"
+            android:visibility="@{archive.isPopular? View.GONE : View.VISIBLE}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toEndOf="@+id/ivGalleryThumb" />
 
-        <Button
+        <com.google.android.material.button.MaterialButton
             android:id="@+id/btnShare"
             style="@style/Widget.MaterialComponents.Button.TextButton.Icon"
             android:layout_width="36dp"
             android:layout_height="36dp"
             android:layout_marginEnd="4dp"
             android:layout_marginBottom="2dp"
-            app:icon="@drawable/ic_link"
+            app:icon="@drawable/ic_link_white"
+            app:iconTint="@{archive.isPopular? @color/middleGrey : @color/white}"
             app:iconSize="24dp"
-            app:iconTint="@color/white"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent" />
